+++ src/Execution/ContextFactory.php
@@ -2,15 +2,18 @@
 
 namespace Nuwave\Lighthouse\Execution;
 
+use Illuminate\Container\Container;
 use Illuminate\Http\Request;
-use Nuwave\Lighthouse\Schema\Context;
+use Nuwave\Lighthouse\Schema\HttpGraphQLContext;
+use Nuwave\Lighthouse\Schema\UserGraphQLContext;
 use Nuwave\Lighthouse\Support\Contracts\CreatesContext;
-use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
 
 class ContextFactory implements CreatesContext
 {
-    public function generate(Request $request): GraphQLContext
+    public function generate(?Request $request): HttpGraphQLContext
     {
-        return new Context($request);
+        return $request
+            ? new HttpGraphQLContext($request)
+            : Container::getInstance()->make(UserGraphQLContext::class);
     }
 }