+++ src/Support/Contracts/GraphQLContext.php
@@ -6,7 +6,7 @@ use Illuminate\Contracts\Auth\Authenticatable;
 use Illuminate\Http\Request;
 
 /**
- * Singleton accesible in all resolvers.
+ * Singleton accessible in all resolvers.
  *
  * @api
  */
@@ -14,16 +14,20 @@ interface GraphQLContext
 {
     /**
      * Get an instance of the authenticated user.
+     *
+     * May be null since some fields may be accessible without authentication.
      */
     public function user(): ?Authenticatable;
-
-    /**
-     * Get an instance of the current HTTP request.
-     */
-    public function request(): Request;
 
     /**
      * Set the authenticated user.
      */
     public function setUser(?Authenticatable $user): void;
+
+    /**
+     * Get an instance of the current HTTP request.
+     *
+     * May be null if GraphQL is run outside the context of an HTTP request.
+     */
+    public function request(): ?Request;
 }