+++ src/Support/Http/GraphQLController.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Nuwave\Lighthouse\Support\Http\Controllers;
+namespace Nuwave\Lighthouse\Support\Http;
 
 use Illuminate\Contracts\Events\Dispatcher as EventsDispatcher;
 use Illuminate\Http\Request;