+++ tests/TestsSerialization.php
@@ -28,13 +28,13 @@ trait TestsSerialization
                         return new User();
                     }
 
+                    public function setUser(?Authenticatable $user): void
+                    {
+                    }
+
                     public function request(): Request
                     {
                         return new Request();
-                    }
-
-                    public function setUser(?Authenticatable $user): void
-                    {
                     }
                 };
             }