+++ tests/Unit/Schema/Execution/CreatesContextTest.php
@@ -8,16 +8,16 @@ use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
 use Tests\TestCase;
 use Tests\Unit\Schema\Execution\Fixtures\FooContext;
 
-final class ContextFactoryTest extends TestCase
+final class CreatesContextTest extends TestCase
 {
     protected function getEnvironmentSetUp($app): void
     {
         parent::getEnvironmentSetUp($app);
 
         $app->singleton(CreatesContext::class, static fn (): CreatesContext => new class() implements CreatesContext {
-            public function generate(Request $request): GraphQLContext
+            public function generate(?Request $request): GraphQLContext
             {
-                return new FooContext($request);
+                return new FooContext();
             }
         });
     }