+++ CHANGELOG.md
@@ -40,6 +40,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Register all non-user namespaces through `RegisterDirectiveNamespaces` event https://github.com/nuwave/lighthouse/pull/2337
 - Guard default configuration options in `lighthouse.php` allows multiple guards https://github.com/nuwave/lighthouse/pull/2345
 - Make `@auth` and `@whereAuth` support multiple guards https://github.com/nuwave/lighthouse/pull/2345
+- Make `Illuminate\Http\Request` optional in `Nuwave\Lighthouse\Support\Contracts\GraphQLContext` https://github.com/nuwave/lighthouse/pull/2353
 
 ### Fixed
 
@@ -84,6 +85,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Remove schema caching v1 https://github.com/nuwave/lighthouse/pull/2321
 - Do not register `ScoutServiceProvider` by default
 - Remove alias/Facade `graphql`
+- Remove subscriptions version 1
 
 ## v5.71.0